classify to return float32

+++ xrspatial/classify.py
@@ -243,7 +243,7 @@ def reclassify(agg: xr.DataArray,
         >>> agg_da = xr.DataArray(data_da, name='agg_da')
         >>> print(agg_da)
         <xarray.DataArray 'agg_da' (dim_0: 4, dim_1: 5)>
-        dask.array<array, shape=(4, 5), dtype=float64, chunksize=(3, 3), chunktype=numpy.ndarray>
+        dask.array<array, shape=(4, 5), dtype=float32, chunksize=(3, 3), chunktype=numpy.ndarray>
         Dimensions without coordinates: dim_0, dim_1
         >>> agg_reclassify_da = reclassify(agg_da, bins=bins, new_values=new_values)  # noqa
         >>> print(agg_reclassify_da)
@@ -402,12 +402,12 @@ def quantile(agg: xr.DataArray,
 def _run_numpy_jenks_matrices(data, n_classes):
     n_data = data.shape[0]
     lower_class_limits = np.zeros(
-        (n_data + 1, n_classes + 1), dtype=np.float64
+        (n_data + 1, n_classes + 1), dtype=np.float32
     )
     lower_class_limits[1, 1:n_classes + 1] = 1.0
 
     var_combinations = np.zeros(
-        (n_data + 1, n_classes + 1), dtype=np.float64
+        (n_data + 1, n_classes + 1), dtype=np.float32
     )
     var_combinations[2:n_data + 1, 1:n_classes + 1] = np.inf
 
@@ -423,7 +423,7 @@ def _run_numpy_jenks_matrices(data, n_classes):
             lower_class_limit = l - m
             i4 = lower_class_limit - 1
 
-            val = np.float64(data[i4])
+            val = np.float32(data[i4])
 
             # here we're estimating variance for each potential classing
             # of the data, for each potential number of classes. `w`
@@ -521,7 +521,7 @@ def _run_jenks(data, n_classes, module):
         lower_class_limits, _ = _run_cupy_jenks_matrices(data, n_classes)
 
     k = data.shape[0]
-    kclass = np.zeros(n_classes + 1, dtype=np.float64)
+    kclass = np.zeros(n_classes + 1, dtype=np.float32)
     kclass[0] = data[0]
     kclass[-1] = data[-1]
     count_num = n_classes
